+++ packages/cli/src/commands/__tests__/mock.spec.ts
@@ -1,5 +1,9 @@
+import * as yargs from 'yargs';
 import { createMultiProcessPrism, createSingleProcessPrism } from '../../util/createServer';
-import Mock from '../mock';
+import mockCommand from '../mock';
+
+const parser = yargs.command(mockCommand);
+
 jest.mock('../../util/createServer');
 
 describe('mock command', () => {
@@ -8,7 +12,10 @@ describe('mock command', () => {
     (createMultiProcessPrism as jest.Mock).mockClear();
   });
   test('starts mock server', async () => {
-    await Mock.run(['/path/to']);
+    await new Promise(resolve => {
+      parser.parse('mock /path/to', () => resolve());
+    });
+
     expect(createMultiProcessPrism).not.toHaveBeenCalled();
     expect(createSingleProcessPrism).toHaveBeenLastCalledWith({
       spec: '/path/to',
@@ -19,7 +26,10 @@ describe('mock command', () => {
   });
 
   test('starts mock server on custom port', async () => {
-    await Mock.run(['-p', '666', '/path/to']);
+    await new Promise(resolve => {
+      parser.parse('mock /path/to -p 666', () => resolve());
+    });
+
     expect(createMultiProcessPrism).not.toHaveBeenCalled();
     expect(createSingleProcessPrism).toHaveBeenLastCalledWith({
       spec: '/path/to',
@@ -30,7 +40,10 @@ describe('mock command', () => {
   });
 
   test('starts mock server on custom host', async () => {
-    await Mock.run(['-h', '0.0.0.0', '/path/to']);
+    await new Promise(resolve => {
+      parser.parse('mock /path/to -h 0.0.0.0', () => resolve());
+    });
+
     expect(createMultiProcessPrism).not.toHaveBeenCalled();
     expect(createSingleProcessPrism).toHaveBeenLastCalledWith({
       spec: '/path/to',
@@ -41,7 +54,10 @@ describe('mock command', () => {
   });
 
   test('starts mock server on custom host and port', async () => {
-    await Mock.run(['-p', '666', '-h', '0.0.0.0', '/path/to']);
+    await new Promise(resolve => {
+      parser.parse('mock /path/to -p 666 -h 0.0.0.0', () => resolve());
+    });
+
     expect(createMultiProcessPrism).not.toHaveBeenCalled();
     expect(createSingleProcessPrism).toHaveBeenLastCalledWith({
       spec: '/path/to',
@@ -52,13 +68,16 @@ describe('mock command', () => {
   });
 
   test('starts mock server with multiprocess option ', async () => {
-    await Mock.run(['-p', '666', '-m', '-h', '0.0.0.0', '/path/to']);
+    await new Promise(resolve => {
+      parser.parse('mock /path/to -m -h 0.0.0.0', () => resolve());
+    });
+
     expect(createSingleProcessPrism).not.toHaveBeenCalled();
     expect(createMultiProcessPrism).toHaveBeenLastCalledWith({
       spec: '/path/to',
       dynamic: false,
       host: '0.0.0.0',
-      port: 666,
+      port: 4010,
     });
   });
 });