+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Prism is now returning the example according to the `Accept` header for OAS2 documents
 - Prism is now `404` when requesting a status code that does not exist in the selected response.
 
+## Changed
+
+- Prism is now using `yargs` for its command line interface, replacing oclif.
+
 # 3.0.1 (2019-07-16)
 
 ## Fixed